+++ modules/apps/foundation/portal-scheduler/portal-scheduler/src/main/java/com/liferay/portal/scheduler/configuration/SchedulerEngineHelperConfiguration.java
@@ -20,7 +20,10 @@ import com.liferay.portal.configuration.metatype.annotations.ExtendedObjectClass
 
 /**
  * @author Michael C. Han
+ * @deprecated As of 3.1.0, replaced by {@link
+ *              com.liferay.portal.scheduler.internal.configuration.SchedulerEngineHelperConfiguration}
  */
+@Deprecated
 @ExtendedObjectClassDefinition(category = "foundation")
 @Meta.OCD(
 	id = "com.liferay.portal.scheduler.configuration.SchedulerEngineHelperConfiguration",