+++ modules/sdk/gradle-plugins-css-builder/src/main/java/com/liferay/gradle/plugins/css/builder/BuildCSSTask.java
@@ -55,6 +55,7 @@ public class BuildCSSTask extends JavaExec {
 		setDefaultCharacterEncoding(StandardCharsets.UTF_8.toString());
 		setDirNames("/");
 		setMain("com.liferay.css.builder.CSSBuilder");
+		systemProperty("sass.compiler.jni.clean.temp.dir", true);
 	}
 
 	public BuildCSSTask dirNames(Iterable<Object> dirNames) {