+++ modules/util/source-formatter/src/main/java/com/liferay/source/formatter/checks/JavaMultiPlusConcatCheck.java
@@ -21,6 +21,9 @@ import com.liferay.portal.kernel.util.StringUtil;
 import com.liferay.portal.kernel.util.Validator;
 import com.liferay.portal.tools.ToolsUtil;
 import com.liferay.source.formatter.parser.JavaTerm;
+import com.liferay.source.formatter.util.FileUtil;
+
+import java.io.File;
 
 import java.util.ArrayList;
 import java.util.List;
@@ -37,34 +40,33 @@ public class JavaMultiPlusConcatCheck extends BaseJavaTermCheck {
 
 	@Override
 	protected String doProcess(
-		String fileName, String absolutePath, JavaTerm javaTerm,
-		String fileContent) {
+			String fileName, String absolutePath, JavaTerm javaTerm,
+			String fileContent)
+		throws Exception {
 
-		if (isExcludedPath(RUN_OUTSIDE_PORTAL_EXCLUDES, absolutePath)) {
+		if (isExcludedPath(RUN_OUTSIDE_PORTAL_EXCLUDES, absolutePath) ||
+			absolutePath.contains("/modules/private/apps/") ||
+			absolutePath.contains("/test/") ||
+			absolutePath.contains("/testIntegration/")) {
+
 			return javaTerm.getContent();
 		}
 
-		if (absolutePath.contains("/portal-impl/") ||
-			absolutePath.contains("/portal-kernel/") ||
-			absolutePath.contains("/portal-test/") ||
-			absolutePath.contains("/portal-test-integration/") ||
-			absolutePath.contains("/util-bridges/") ||
-			absolutePath.contains("/util-java/") ||
-			absolutePath.contains("/util-taglib/")) {
-
-			_checkConcat(fileName, javaTerm.getContent(), fileContent);
-		}
+		_checkConcat(
+			fileName, absolutePath, javaTerm.getContent(), fileContent);
 
 		return javaTerm.getContent();
 	}
 
 	@Override
 	protected String[] getCheckableJavaTermNames() {
-		return new String[] {JAVA_CONSTRUCTOR, JAVA_METHOD};
+		return new String[] {JAVA_CONSTRUCTOR, JAVA_METHOD, JAVA_VARIABLE};
 	}
 
 	private void _checkConcat(
-		String fileName, String javaTermContent, String fileContent) {
+			String fileName, String absolutePath, String javaTermContent,
+			String fileContent)
+		throws Exception {
 
 		int x = -1;
 
@@ -115,6 +117,13 @@ public class JavaMultiPlusConcatCheck extends BaseJavaTermCheck {
 			if ((parts.size() > 3) &&
 				_containsStringVariable(fileContent, parts)) {
 
+				if (absolutePath.contains("/modules/") &&
+					!absolutePath.contains("/modules/apps/") &&
+					!_hasKernelOrPetraStringDependency(fileName)) {
+
+					return;
+				}
+
 				int pos = fileContent.indexOf(plusStatement);
 
 				addMessage(
@@ -138,7 +147,7 @@ public class JavaMultiPlusConcatCheck extends BaseJavaTermCheck {
 				part.startsWith("File.separator") ||
 				part.startsWith("StringPool.") ||
 				(Validator.isVariableName(part) &&
-				 content.matches("(?s).+\\sString\\s+" + part + "\\w.+"))) {
+				 content.matches("(?s).+\\sString\\s+" + part + "\\W.+"))) {
 
 				return true;
 			}
@@ -225,6 +234,39 @@ public class JavaMultiPlusConcatCheck extends BaseJavaTermCheck {
 					return i + 1;
 				}
 			}
+		}
+	}
+
+	private boolean _hasKernelOrPetraStringDependency(String fileName)
+		throws Exception {
+
+		int x = fileName.length();
+
+		while (true) {
+			x = fileName.lastIndexOf("/", x - 1);
+
+			if (x == -1) {
+				return false;
+			}
+
+			String buildGradleFileName =
+				fileName.substring(0, x + 1) + "build.gradle";
+
+			File file = new File(buildGradleFileName);
+
+			if (!file.exists()) {
+				continue;
+			}
+
+			String content = FileUtil.read(file);
+
+			if (content.contains("name: \"com.liferay.petra.string\"") ||
+				content.contains("name: \"com.liferay.portal.kernel\"")) {
+
+				return true;
+			}
+
+			return false;
 		}
 	}